Fixed conversion warnings in tests.
The signature of `flags.load` takes an `int`, but
`arraySize` (correctly) returns a `size_t`. We don't want to change
`flags.load` to take a `size_t` because its signature is `int argc,
char** argv`, and so we may be sending it `int` in other places.

As all the test arrays are statically sized, we know that we're not
overflowing an `int`, so it is safe to just `static_cast<int>` here.

Review: https://reviews.apache.org/r/64732

+++ 3rdparty/stout/tests/subcommand_tests.cpp
@@ -113,7 +113,7 @@ TEST(SubcommandTest, Flags)
 
   // Construct the command line arguments.
   vector<string> _argv = getArgv(flags);
-  int argc = _argv.size() + 2;
+  int argc = static_cast<int>(_argv.size()) + 2;
   char** argv = new char*[argc];
   argv[0] = (char*) "command";
   argv[1] = (char*) "subcommand";
